+++ src/main/java/org/truffleruby/builtins/CoreMethodNodeManager.java
@@ -267,8 +267,7 @@ public class CoreMethodNodeManager {
             if (alwaysInlined) {
                 callTarget = callTargetFactory.apply(sharedMethodInfo);
                 callTargetSupplier = null;
-                final RubyRootNode rootNode = RubyRootNode.of(callTarget);
-                alwaysInlinedNodeFactory = ((ReRaiseInlinedExceptionNode) rootNode.getBody()).nodeFactory;
+                alwaysInlinedNodeFactory = RubyRootNode.of(callTarget).getAlwaysInlinedNodeFactory();
             } else {
                 if (context.getLanguageSlow().options.LAZY_CALLTARGETS) {
                     callTarget = null;