+++ src/main/java/org/truffleruby/language/RubyCoreMethodRootNode.java
@@ -18,16 +18,11 @@ import org.truffleruby.language.methods.Split;
 import org.truffleruby.language.methods.TranslateExceptionNode;
 
 import com.oracle.truffle.api.CompilerDirectives;
-import com.oracle.truffle.api.RootCallTarget;
 import com.oracle.truffle.api.frame.FrameDescriptor;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.source.SourceSection;
 
 public class RubyCoreMethodRootNode extends RubyCheckArityRootNode {
-
-    public static RubyCoreMethodRootNode of(RootCallTarget callTarget) {
-        return (RubyCoreMethodRootNode) callTarget.getRootNode();
-    }
 
     @Child private TranslateExceptionNode translateExceptionNode;