+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Bug fixes:
 * `BasicSocket#*_nonblock(exception: false)` now only return `:wait_readable/:wait_writable` for `EAGAIN`/`EWOULDBLOCK` like MRI (#2400).
 * Fix issue with `strspn` used in the `date` C extension compiled as a macro on older glibc and then missing the `__strspn_c1` symbol on newer glibc (#2406).
 * Fix constant lookup when loading the same file multiple times (#2408).
+* Fix handling of `break`, `next` and `redo` in `define_method(name, &block)` methods (#2418).
 
 Compatibility: